point3c compliance
